test: 도메인 util 함수 테스트 케이스 분리
